requirements/local.txt
django-debug-toolbar==1.4
django-extensions
ipython
